requirements.txt
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
Cython==0.29.20
decorator==4.4.2
docopt==0.6.2
easydict==1.9
gluoncv==0.7.0
graphviz==0.8.4
idna==2.10
imageio==2.8.0
imutils==0.5.3
joblib==0.15.1
kiwisolver==1.2.0
matplotlib==3.2.2
mxnet-cu100==1.5.1.post0
networkx==2.4
numpy==1.19.0
opencv-python==4.2.0.34
Pillow==7.2.0
pipreqs==0.4.10
portalocker==1.7.0
PyMySQL==0.9.3
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.24.0
scikit-image==0.17.2
scikit-learn==0.22
scipy==1.5.0
six==1.15.0
tifffile==2020.6.3
tqdm==4.47.0
urllib3==1.25.9
yarg==0.1.9
sanic
Sanic-Cors
ujson